Learning Application Development/backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# For SQLite (no additional driver needed)
# For PostgreSQL (uncomment if needed later)
# psycopg2-binary==2.9.7

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Background tasks
celery==5.3.4
redis==5.0.1

# AI and ML
openai==1.3.5
requests==2.31.0
beautifulsoup4==4.12.2
pypdf2==3.0.1
pdfplumber==0.10.3
numpy==1.24.3
scikit-learn==1.3.2

# YouTube processing
yt-dlp==2023.11.16

# Vector database (optional)
# pinecone-client==2.2.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
